suporte

+++ models.py
@@ -1,6 +1,6 @@
 from __future__ import annotations
 from sqlalchemy.orm import DeclarativeBase, Mapped, mapped_column
-from sqlalchemy import String, Integer, Boolean, DateTime, Text, Time
+from sqlalchemy import String, Integer, BigInteger, Boolean, DateTime, Text, Time
 from datetime import datetime, timezone, time as dtime
 
 from typing import Optional
@@ -17,7 +17,7 @@ class Config(Base):
 class User(Base):
     __tablename__ = "users"
     id: Mapped[int] = mapped_column(Integer, primary_key=True, autoincrement=True)
-    tg_id: Mapped[int] = mapped_column(Integer, unique=True, index=True)
+    tg_id: Mapped[int] = mapped_column(BigInteger, unique=True, index=True)
     username: Mapped[str] = mapped_column(String(64), default="", nullable=False)
     is_vip: Mapped[bool] = mapped_column(Boolean, default=False)
     vip_until: Mapped[Optional[datetime]] = mapped_column(DateTime(timezone=True), nullable=True)
@@ -29,7 +29,7 @@ class PendingNotification(Base):
     """Mensagens que não puderam ser enviadas (usuário não iniciou conversa)"""
     __tablename__ = "pending_notifications"
     id: Mapped[int] = mapped_column(Integer, primary_key=True, autoincrement=True)
-    user_id: Mapped[int] = mapped_column(Integer, index=True, nullable=False)
+    user_id: Mapped[int] = mapped_column(BigInteger, index=True, nullable=False)
     username: Mapped[Optional[str]] = mapped_column(String(64), nullable=True)
     message: Mapped[str] = mapped_column(Text, nullable=False)
     created_at: Mapped[datetime] = mapped_column(
@@ -42,7 +42,7 @@ class SupportTicket(Base):
     """Tickets de suporte dos usuários"""
     __tablename__ = "support_tickets"
     id: Mapped[int] = mapped_column(Integer, primary_key=True, autoincrement=True)
-    user_id: Mapped[int] = mapped_column(Integer, index=True, nullable=False)
+    user_id: Mapped[int] = mapped_column(BigInteger, index=True, nullable=False)
     username: Mapped[Optional[str]] = mapped_column(String(64), nullable=True)
     first_name: Mapped[Optional[str]] = mapped_column(String(120), nullable=True)
     description: Mapped[str] = mapped_column(Text, nullable=False)
@@ -58,7 +58,7 @@ class MemberLog(Base):
     """Log de entrada/saída de membros no grupo VIP"""
     __tablename__ = "member_logs"
     id: Mapped[int] = mapped_column(Integer, primary_key=True, autoincrement=True)
-    user_id: Mapped[int] = mapped_column(Integer, index=True, nullable=False)
+    user_id: Mapped[int] = mapped_column(BigInteger, index=True, nullable=False)
     username: Mapped[Optional[str]] = mapped_column(String(64), nullable=True)
     first_name: Mapped[Optional[str]] = mapped_column(String(120), nullable=True)
     action: Mapped[str] = mapped_column(String(20), nullable=False)  # "joined", "left", "removed"
